Fix file names, align with default standard
