--- src/FilterInterface.ts
@@ -1,5 +1,0 @@
-import { FilterBuilderInterface } from "./FilterBuilderInterface";
-
-export interface FilterInterface {
-  build(builder: FilterBuilderInterface): object;
-}
