--- src/adapters/EngineInterface.ts
@@ -1,6 +1,0 @@
-import { EngineResponse } from "../EngineResponse";
-import { EngineRequest } from "../EngineRequest";
-
-export interface EngineInterface {
-  execute(request: EngineRequest): EngineResponse;
-}
