+++ src/adapters/engine-interface.ts
@@ -1,0 +1,6 @@
+import { EngineResponse } from "../engine-response";
+import { EngineRequest } from "../engine-request";
+
+export interface EngineInterface {
+  execute(request: EngineRequest): EngineResponse;
+}
