+++ src/fIlter-builder-interface.ts
@@ -1,7 +1,7 @@
-import { FilterExpression } from "./filters/FilterExpression";
-import { FilterList } from "./filters/FilterList";
+import { FilterExpression } from "./filters/filter-expression";
+import { FilterList } from "./filters/filter-list";
 
-export interface FilterBuilderInterface {
+export interface FIlterBuilderInterface {
   buildFilter(filter: FilterList): object;
   validate(expression: FilterExpression): boolean;
   buildFilterExpression(expression: FilterExpression): object;
