+++ src/filter-Interface.ts
@@ -1,0 +1,5 @@
+import { FIlterBuilderInterface } from "./fIlter-builder-interface";
+
+export interface FilterInterface {
+  build(builder: FIlterBuilderInterface): object;
+}
