+++ src/index.ts
@@ -1,12 +1,12 @@
-export * from "./Configuration";
-export * from "./filters/FilterBuilder";
-export * from "./adapters/AWSShellEngineAdapter";
-export * from "./adapters/EngineInterface";
+export * from "./configuration";
+export * from "./filters/filter-builder";
+export * from "./adapters/aws-shell-engine-adapter";
+export * from "./adapters/engine-interface";
 
 export * from "./domain/resource-type";
 export * from "./domain/types/ebs";
-export * from "./filters/Criteria";
-export * from "./filters/C7nFilterBuilder";
+export * from "./filters/criteria";
+export * from "./filters/c7n-filter-builder";
 
 export * from "./responses/detached-volumes-response";
 export * from "./responses/ebs-garbage-item";
